Credentials refactoring to support other GCP services.
