+++ vision_client.py
@@ -1,8 +1,8 @@
 from google.cloud import vision
 from google.cloud.vision import types
-from google.oauth2 import service_account
+
 import logger
-import constants
+import gcp_credentials
 
 
 def send_image(content):
@@ -36,23 +36,9 @@ def send_image(content):
 
 
 def _get_client():
-    keyfile_dict = _get_keyfile_dict()
-    scope = [constants.GCP_SCOPE]
-    creds = service_account.Credentials.from_service_account_info(keyfile_dict).with_scopes(scope)
-
+    creds = gcp_credentials.get_credentials()
     client = vision.ImageAnnotatorClient(credentials=creds)
     return client
-
-
-def _get_keyfile_dict():
-    keyfile_dict = {}
-    keyfile_dict['type'] = constants.GCP_TYPE
-    keyfile_dict['client_email'] = constants.GCP_CLIENT_EMAIL
-    keyfile_dict['private_key'] = constants.GCP_PRIVATE_KEY
-    keyfile_dict['private_key_id'] = constants.GCP_PRIVATE_KEY_ID
-    keyfile_dict['client_id'] = constants.GCP_CLIENT_ID
-    keyfile_dict['token_uri'] = constants.GCP_TOKEN_URI
-    return keyfile_dict
 
 
 # Instantiates a client
